requirements.txt
Django~=4.0.4
requests~=2.26.0
djangorestframework~=3.13.1
django-celery-beat
drf_yasg
celery~=5.2.7